UI Refactor, new JavaFX UI, Initial Preferences support
Chart construction extracted to a dedicated class (extends JFreeChart).
Added a very simplistic mechanism for flight detection and duration.

Swing UI refactored - classes moved to new swing package

JavaFX exercise. Build the application using JavaFX's Chart classes is
complicated because they can't handle too many points. Fortunately
there's FXGraphics2D to bridge swing and JavaFX.

Initial support to application preferences, allowing user defined
colors, limits, and other algorithm adjustments.

Altimeter file reader now sets current timestamp for each sample and
minimum/maximum values found in each session.

+++ src/main/java/org/psicover/altimeter/bean/SampleRate.java
@@ -19,7 +19,7 @@ public enum SampleRate implements Serializable {
 		return (1<<ordinal());
 	}
 	
-	public int duration(int numSamples) {
+	public double duration(double numSamples) {
 		return (numSamples/samplesPerSecond());
 	}
